+++ errs/errs.go
@@ -10,4 +10,5 @@ var (
 	ErrNotFound = eris.New("not found")
 	// ErrNoValue is the error for when no value is found.
 	ErrNoValue = eris.New("no value")
+	ErrNever = eris.New("never ran")
 )