+++ state/state.go
@@ -4,6 +4,7 @@ package state
 import (
 	"context"
 	"fmt"
+	"github.com/hjertnes/roam/errs"
 
 	"github.com/hjertnes/roam/configuration"
 	dal2 "github.com/hjertnes/roam/dal"
@@ -21,8 +22,16 @@ type State struct {
 	Dal  dal2.Dal
 }
 
+
 // New is the constructor.
+
+func NewWithoutStatusCheck(path string) (*State, error) {
+	return _new(false, path)
+}
 func New(path string) (*State, error) {
+	return _new(true, path)
+}
+func _new(check bool, path string) (*State, error) {
 	conf, err := configuration.ReadConfigurationFile(fmt.Sprintf("%s/.config/config.yaml", path))
 	if err != nil {
 		return nil, eris.Wrap(err, "failed to get config")
@@ -36,6 +45,20 @@ func New(path string) (*State, error) {
 	}
 
 	dal := dal2.New(ctx, pxp, path)
+
+	if check{
+		s, err := dal.WasLastSyncSuccessful()
+		if err != nil {
+			if eris.Is(err, errs.ErrNever){
+				fmt.Println("Sync have never been done")
+			} else {
+				return nil, eris.Wrap(err, "failed to check last sync")
+			}
+
+		} else if !s{
+			fmt.Println("Last sync failed. Run diagnostic to see why")
+		}
+	}
 
 	return &State{
 		Conn: pxp,